requirements.txt
chromadb==0.5.3
sentence-transformers==3.0.1
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
litellm==1.45.0
httpx==0.27.2
pillow==10.4.0
python-multipart==0.0.9
